Add state tests to related methods
