+++ test/route-hash-navigation.js
