+++ test/route-methods.js
@@ -1,10 +1,7 @@
 const PushStateTree = require('../src/pushStateTree');
-import cleanHistoryAPI from './helper/cleanHistoryAPI';
-const _ = require('underscore');
 
 describe('PushStateTree methods', function() {
   let pst;
-  cleanHistoryAPI();
 
   beforeEach(() => {
     pst = new PushStateTree();
@@ -27,20 +24,6 @@ describe('PushStateTree methods', function() {
       'replace',
       'navigate'
     ].forEach(method => expect(methods).to.contain(method));
-  });
-
-  it('should set the state when use pushState method', () => {
-    let uniqueState = {uniqueState: _.uniqueId('state')};
-    pst.pushState(uniqueState, '', _.uniqueId('new_url'));
-
-    assert.deepEqual(pst.state, uniqueState);
-  });
-
-  it('should set the state to null when use navigate method', () => {
-    let uniqueState = {uniqueState: _.uniqueId('state')};
-    pst.navigate(_.uniqueId('new_url'));
-
-    assert.isNull(pst.state, 'the state is empty');
   });
 
 });
